Coverage: Add an option to set the minimal access level of included symbols.
